telegram-bot/pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
]
